reseni_konflikti_kod_rejtinga

+++ BookingAppTeam05/src/main/java/com/example/BookingAppTeam05/service/DeleteAccountService.java
@@ -55,6 +55,9 @@ public class DeleteAccountService {
             if (deleteAccountRequest == null)
                 return "Can't find delete account request with id: " + d.getId();
 
+            if (deleteAccountRequest.isProcessed())
+                return "This request is already processed";
+
             deleteAccountRequest.setProcessed(true);
             deleteAccountRequest.setAccepted(d.isAccepted());
             deleteAccountRepository.delete(deleteAccountRequest);
